+++ tests/test-wiggle.R
@@ -38,5 +38,40 @@ stopifnot(isTRUE(all.equal(
 )))
 
 ################################################################################
+#                              Count Sign Changes                              #
+
+stop("DEV WORK - PICK UP HERE")
+fprime(x) |>
+  sign() |>
+  diff() |>
+  abs() |>
+  sum() / 2L
+
+
+fdoubleprime(x) |>
+  sign() |>
+  diff() |>
+  abs() |>
+  sum() / 2L
+
+sign_changes( object = cp(bmat, theta), lower = -3, upper = 5, derivative = 1L)
+sign_changes( object = cp(bmat, theta), lower = -3, upper = 5, derivative = 2L)
+
+plot(x, bmatD2 %*% theta)
+
+object <- cp(bmat, theta)
+  f <- function(x) {
+    bsplineD(x,
+              iknots = object$iknots,
+              bknots = object$bknots,
+              order  = object$order,
+              derivative = 2L) %*%
+     matrix(object$cp$theta, ncol = 1)
+  }
+
+x2 <- seq(min(bknots), max(bknots), length.out = 100000)
+points(x2, f(x2), col = "red", type = "l")
+
+################################################################################
 #                                 End of File                                  #
 ################################################################################